refactor(types): the validator scripts stop faking an Env by assertion
24 casts across five validator scripts, all the same shape:
`createLocalArtifactEnv() as unknown as Env` and its multi-line sibling,
each script inventing its own way to hand a Worker handler an env.

The suites had already solved this: `mockEnv` in tests/row-type.ts wraps
the one honest assertion -- a stub standing in for live platform
bindings a unit process cannot construct -- behind a named helper. The
scripts drive the SAME handlers and were doing it by hand.

`scripts/lib/worker-env.ts` (moved from tests/helpers, which now
re-exports it) serves both. Its parameter is keyed on the real env with
`unknown` values, so a binding NAME the Worker does not have is still an
error while a hand-rolled stub is allowed.

I first tried making `createLocalArtifactEnv` RETURN an `ApiWorkerEnv`,
which is the more obviously correct fix. It is not the better one: the
suites build an env and then mutate it (`env.DATA_API = { fetch }`), and
a real `Fetcher` also declares `connect`, so a typed return pushed
platform-type strictness into ~30 fixtures to remove 24 casts from
scripts. Wrapping at the call sites gets the same result with no test
churn.

That attempt did surface four more fixtures still setting the retired
`METAGRAPH_HEALTH_DB`, and a tripwire -- "rejects unsupported query
parameters before the store" -- spying on it, which therefore cannot
fire. Worth fixing on their own; not in this commit.

scripts: 54 -> 30.

Verified: tsc clean, eslint clean (uncached), prettier clean,
validate:api green (it drives the real handlers through this change),
plus double-assertions / types / boundary-casts / unreferenced-exports.
