+++ scripts/validate-api.ts
@@ -22,6 +22,7 @@ import { buildAccountIdentity } from "../src/account-identity.ts";
 import { blockEmissionForIssuance } from "../src/block-emission.ts";
 import { taoToRao } from "../src/emission-decomposition.ts";
 import {} from "../workers/request-params.ts";
+import { apiEnv } from "./lib/worker-env.ts";
 
 // OpenAPI document + Worker response bodies are dynamic JSON read only for
 // assertion purposes -- never trusted for control flow. Mirrors the
@@ -3027,7 +3028,7 @@ for (const [route, assertion, options = {}] of checks) {
             body: JSON.stringify(options.body),
           },
     ),
-    env as unknown as Env,
+    apiEnv(env),
     {},
   );
   assert.equal(
@@ -3231,7 +3232,7 @@ for (const [route, assertion, options = {}] of checks) {
     });
     const response = await handleRequest(
       new Request(`https://metagraph.sh${route}`),
-      forwardingEnv as unknown as Env,
+      apiEnv(forwardingEnv),
       {},
     );
     assert.equal(response.status, 200, `${flag} ${route}: expected 200`);
@@ -3313,7 +3314,7 @@ for (const [route, assertion, options = {}] of checks) {
     });
     await handleRequest(
       new Request(`https://metagraph.sh${route}`),
-      sweptEnv as unknown as Env,
+      apiEnv(sweptEnv),
       {},
     );
     assert.equal(
@@ -3335,7 +3336,7 @@ const paginated = await handleRequest(
   new Request(
     "https://metagraph.sh/api/v1/subnets?limit=2&sort=netuid&order=desc",
   ),
-  env as unknown as Env,
+  apiEnv(env),
   {},
 );
 const paginatedBody = (await paginated.json()) as Row;
@@ -3358,7 +3359,7 @@ for (const route of [
 ]) {
   const response = await handleRequest(
     new Request(`https://metagraph.sh${route}`),
-    env as unknown as Env,
+    apiEnv(env),
     {},
   );
   assert.equal(response.status, 400, `${route}: expected invalid query`);
@@ -3371,7 +3372,7 @@ for (const route of [
 
 const etagSource = await handleRequest(
   new Request("https://metagraph.sh/api/v1/subnets/7"),
-  env as unknown as Env,
+  apiEnv(env),
   {},
 );
 const cached = await handleRequest(
@@ -3380,14 +3381,14 @@ const cached = await handleRequest(
       "if-none-match": etagSource.headers.get("etag")!,
     },
   }),
-  env as unknown as Env,
+  apiEnv(env),
   {},
 );
 assert.equal(cached.status, 304, "matching ETag should return 304");
 
 const missing = await handleRequest(
   new Request("https://metagraph.sh/api/v1/subnets/9999"),
-  env as unknown as Env,
+  apiEnv(env),
   {},
 );
 assert.equal(missing.status, 404, "missing subnet should return 404");
@@ -3399,7 +3400,7 @@ assert.equal(
 
 const proxy = await handleRequest(
   new Request("https://metagraph.sh/rpc/v1/finney", { method: "POST" }),
-  env as unknown as Env,
+  apiEnv(env),
   {},
 );
 assert.equal(proxy.status, 501, "RPC proxy should be disabled by default");
@@ -3414,10 +3415,10 @@ const blockedRpc = await handleRequest(
       params: [],
     }),
   }),
-  {
+  apiEnv({
     ...env,
     METAGRAPH_ENABLE_RPC_PROXY: "true",
-  } as unknown as Env,
+  }),
   {},
 );
 assert.equal(
@@ -3434,7 +3435,7 @@ const verifyMissing = await handleRequest(
   new Request(
     "https://metagraph.sh/api/v1/surfaces/zzz-not-a-real-surface/verify",
   ),
-  env as unknown as Env,
+  apiEnv(env),
   {},
 );
 assert.equal(
@@ -3450,7 +3451,7 @@ assert.equal(
 
 const r2Fallback = await handleRequest(
   new Request("https://metagraph.sh/api/v1/changelog"),
-  {
+  apiEnv({
     ASSETS: {
       async fetch() {
         return new Response("not found", { status: 404 });
@@ -3477,7 +3478,7 @@ const r2Fallback = await handleRequest(
         };
       },
     },
-  } as unknown as Env,
+  }),
   {},
 );
 assert.equal(
